DocString Added and Single Principle Responsability Error Client Applied.
