+++ src/client/error/APIError.py
@@ -1,0 +1,4 @@
+class APIError(Exception):
+    def __init__(self, status_code, message):
+        self.status_code = status_code
+        super().__init__(message)
