+++ src/client/error/AuthError.py
@@ -1,0 +1,3 @@
+from .APIError import APIError
+
+class AuthError(APIError): pass
