+++ src/client/error/ForbiddenError.py
@@ -1,0 +1,3 @@
+from .APIError import APIError
+
+class ForbiddenError(APIError): pass
