+++ src/client/error/NotFoundError.py
@@ -1,0 +1,3 @@
+from .APIError import APIError
+
+class NotFoundError(APIError): pass
