+++ src/client/error/ValidationError.py
@@ -1,0 +1,3 @@
+from .APIError import APIError
+
+class ValidationError(APIError): pass
