+++ src/client/error/__init__.py
@@ -1,0 +1,5 @@
+from .APIError import APIError
+from .AuthError import AuthError
+from .ForbiddenError import ForbiddenError
+from .NotFoundError import NotFoundError
+from .ValidationError import ValidationError
